backend/requirements.txt
fastapi[all]==0.121.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
asyncpg>=0.30.0
python-jose[cryptography]==3.5.0
bcrypt>=4.0.0
python-multipart==0.0.20
pydantic[email]>=2.1.1,<3.0.0
alembic==1.17.1
python-dotenv==1.2.1
python-keycloak>=2.10.0
minio>=7.2.0
pillow>=10.0.0